add option: DNSdata_adjust_C1C2 = zbrent
I tried this because the surface finder (that finds sigma) failed.
The failure was due to a too large Newton step for C1 that resulted in
a q that was negative everwhere. Then there was no star surface...
But zbrent is less accurate, maybe I need to use it with a smaller tol.