revert changes 3562d8a and 6cb2185 to Sigma extrapolation
We actually always imposed continuity in n^i d/dx^x Sigma, which
is equivalent to imposing it in d/dr Sigma. So there was no bug!
Except in the future we could try imposing it directly in d/dr Sigma
via d/dlam Sigma, which should be cheaper, and could be numerically
different.